# public/swagger/waiting-number-service.yml
swagger: '2.0'
info:
  title: Waiting number service
  description: Manage your waiting numbers
  version: "1.0.0"
host: localhost:3000
schemes:
  - http
basePath: /
produces:
  - application/json
paths:
  /waitingTicket:
    get:
      summary: Get all waiting tickets
      description: Retrieve an array of all waiting tickets.
      responses:
        200:
          description: Array of waiting tickets
          schema:
            type: array
            items:
             $ref: '#/definitions/WaitingTicket'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
  /waitingTicket/newTicket:
    get:
      summary: Get new waiting ticket
      description: Request a new waiting ticket for a client.
      responses:
        201:
          description: The new ticket
          schema:
            $ref: '#/definitions/WaitingTicket'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
  /waitingTicket/nextTicket:
    get:
      summary: Next ticket in line
      description: Get the waiting ticket for the next waiting client.
      responses:
        200:
          description: The ticket
          schema:
            $ref: '#/definitions/WaitingTicket'
        404:
          description: No client waiting
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
  /waitingTicket/{ticketId}/takeIntoProcess:
    put:
      summary: Take a given waiting ticket into process
      description: Takes a waiting ticket identified by the id into process by the provided desk. Only works, if the ticket wasn't processed already.
      responses:
        200:
          description: Ticket taken sucessfully
        400:
          description: Parameters missing
        404:
          description: Ticket not found
        409:
          description: Ticket already in process
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
      parameters:
        - name: ticketId
          in: path
          description: The id of the waiting ticket
          required: true
          type: string
        - name: deskNumber
          in: query
          description: Your desk's number
          required: true
          type: integer
          format: int32
            
definitions:
  WaitingTicket:
    type: object
    properties:
      callNumber:
        type: integer
        format: int32
        description: Ticket number to display on the screen.
      deskNumber:
        type: integer
        format: int32
        description: Desk that was assigned to this waiting ticket. Equals to 0 if the ticket wasn't dispached yet.
  Error:
    type: object
    properties:
      code:
        type: integer
        format: int32
      message:
        type: string